Update BaseEntity
- Create EBlogNotification, EBookingNotification, EAges, TourTemplate
- Update
 + Blog: add relation with Tour, Blog
 + BlogNotification : change type of notificationType to EBlogNotification
 + BookingNotification : change type of notificationType to EBooking
 + Booking : change type of status to EBooking
 + PersonBooking : add ages to manage ages, add isCancel for manage canceled booking
 + Review: add relation with review
 + Tour, TourTemplate : split Tour to 2 tables Tour & TourTemplate, delete isVerify, change date from LocalDateTime to Integer
 + TourPrice: change type to Integer for manage percent price of ages

+++ travel_social/src/main/java/com/app/entity/TourPrice.java
@@ -17,9 +17,9 @@ import java.math.BigDecimal;
 @Getter
 @Setter
 public class TourPrice extends BaseEntity {
-    private BigDecimal adult;
-    private BigDecimal children;
-    private BigDecimal baby;
+    private Integer adult;
+    private Integer children;
+    private Integer baby;
 
     @ManyToOne
     @JoinColumn(name = "TOUR_ID", referencedColumnName = "ID")
